updt moe to have single pg for all depth instead of one pg for each moe layer
