+++ examples/llm/yamls/mosaic_gpt/125m_moedev.yaml
@@ -17,10 +17,11 @@ model:
   n_layers: 12
   mlp_ratio: 4
   moe:
-    num_experts: [1, 2, 4, 8, 12, 16, 20, 24, 28, 32, 36, 40]  # on 4 gpus runs with: 2 experts (ie 2 gpus share an expert) and 8 (ie each gpu has 2 experts)
+    moe_type: tutel
+    num_experts: [1, 2, 4, 8, 12, 16, 20, 24, 28, 32, 36, 40]  # list or int; if list, must be length number of MoE layers
     l_aux_wt: 0.01
     a2a_ffn_overlap_degree: 2
-    use_2dh: false
+    use_2dh: true
     capacity_factor: -2
     gate_noise: 0.01
     init_experts: true
@@ -49,7 +50,6 @@ train_loader:
     shuffle: true
     tokenizer_name: ${tokenizer_name}
     max_seq_len: ${max_seq_len}
-    group_method: concat
     shuffle_seed: ${global_seed}
   drop_last: true
   num_workers: 8
@@ -63,7 +63,6 @@ eval_loader:
     shuffle: false
     tokenizer_name: ${tokenizer_name}
     max_seq_len: ${max_seq_len}
-    group_method: truncate
     shuffle_seed: ${global_seed}
   drop_last: false
   num_workers: 8
@@ -88,7 +87,7 @@ algorithms:
     clipping_type: norm
     clipping_threshold: 1.0
 
-max_duration: 10ba # ~ 2.5B tokens
+max_duration: 16ba # ~ 2.5B tokens
 eval_interval: 500ba
 eval_subset_num_batches: 2
 global_train_batch_size: 32
@@ -115,7 +114,7 @@ console_log_interval: 1ba
 
 callbacks:
   speed_monitor:
-    window_size: 10
+    window_size: 8
   lr_monitor: {}
   memory_monitor: {}
